game scenes refactor
