+++ src/components/Play/games/shared.js
@@ -1,0 +1,4 @@
+export const BG_KEY = "gymfloor";
+export const PLAYER_KEY = "avatar";
+export const LEFT_CHEVRON = 'left_chevron';
+export const PLAYER_SCALE = 0.15;
